Using `usize` to compute file offsets caps us at ~16GB on 32-bit
systems. For example, with 4 KiB pages we can only address up to 1048576
pages; attempting the next page overflows a 32-bit usize and can wrap
the write offset, corrupting data. Switching our I/O APIs and offset
math to u64 avoids this overflow on 32-bit targets

Closes #2791

+++ simulator/runner/file.rs
@@ -150,7 +150,7 @@ impl File for SimulatorFile {
         self.inner.unlock_file()
     }
 
-    fn pread(&self, pos: usize, c: turso_core::Completion) -> Result<turso_core::Completion> {
+    fn pread(&self, pos: u64, c: turso_core::Completion) -> Result<turso_core::Completion> {
         self.nr_pread_calls.set(self.nr_pread_calls.get() + 1);
         if self.fault.get() {
             tracing::debug!("pread fault");
@@ -173,7 +173,7 @@ impl File for SimulatorFile {
 
     fn pwrite(
         &self,
-        pos: usize,
+        pos: u64,
         buffer: Arc<turso_core::Buffer>,
         c: turso_core::Completion,
     ) -> Result<turso_core::Completion> {
@@ -227,7 +227,7 @@ impl File for SimulatorFile {
 
     fn pwritev(
         &self,
-        pos: usize,
+        pos: u64,
         buffers: Vec<Arc<turso_core::Buffer>>,
         c: turso_core::Completion,
     ) -> Result<turso_core::Completion> {
@@ -257,7 +257,7 @@ impl File for SimulatorFile {
         self.inner.size()
     }
 
-    fn truncate(&self, len: usize, c: turso_core::Completion) -> Result<turso_core::Completion> {
+    fn truncate(&self, len: u64, c: turso_core::Completion) -> Result<turso_core::Completion> {
         if self.fault.get() {
             return Err(turso_core::LimboError::InternalError(
                 FAULT_ERROR_MSG.into(),
